[Feat] dashboard complet (Tablette + PC)
