+++ src/routes/routes.tsx
@@ -19,7 +19,6 @@ import HabitProfils from "../pages/admin/statistiques/HabitProfils";
 import ConsultProfils from "../pages/admin/statistiques/ConsultProfilsPage";
 import AppPreparateur from "../pages/preparateur/AppPreparateur";
 import DashboardPrepa from "../pages/preparateur/dashboard/DashboardPrepaPage";
-import DetailsPage from "../pages/preparateur/details/DetailsPage";
 
 const router = createBrowserRouter([
   {
@@ -113,10 +112,6 @@ const router = createBrowserRouter([
               {
                 path: "dashboard",
                 element: <DashboardPrepa />,
-              },
-              {
-                path: "dashboard/:idCommande",
-                element: <DetailsPage />,
               },
             ],
           },
